--create separate module for parameter handling
